Applied Eclipse's cleanup function on all projects with very conservative settings.
git-svn-id: https://svn.jfire.org/svn/main/trunk/JFireScriptingEditor2D@9054 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/jfire/scripting/editor2d/impl/BarcodeDrawComponentImpl.java
@@ -37,7 +37,6 @@ import org.nightlabs.editor2d.DrawComponentContainer;
 import org.nightlabs.editor2d.impl.DrawComponentImpl;
 import org.nightlabs.editor2d.render.BaseRenderer;
 import org.nightlabs.editor2d.render.Renderer;
-import org.nightlabs.editor2d.render.j2d.J2DRenderContext;
 import org.nightlabs.i18n.unit.IUnit;
 import org.nightlabs.i18n.unit.MMUnit;
 import org.nightlabs.i18n.unit.resolution.DPIResolutionUnit;
@@ -292,7 +291,7 @@ implements BarcodeDrawComponent
 		}
 		
 		int resolution = getModelResolution();
-		double factor = ((double)resolution) / 300d;		
+		double factor = (resolution) / 300d;		
 		double scaledWidth = width * factor;
 		if (logger.isDebugEnabled()) {
 			logger.debug("width = " + width);
